Fix pipeline blockers: pg auth, distance_m type, geometry guard
BLOCKER 1: Set pg_hba.conf to trust auth so pike user works when
container runs as root.

RISK 1: Change distance_m in exit_place_links from INTEGER to DOUBLE
PRECISION so float values in OI CSV don't break \copy. Update Rust
PendingPair.air_distance_m to f64 accordingly.

RISK 2: Add Point type guard to places geometry_geojson lat/lon
extraction.

RISK 3: Export osrm_snap_hints from database to CSV before S3 upload
so snap hints are included in releases.

+++ src/exit_poi_linker/reachability/db.rs
@@ -12,7 +12,7 @@ pub(super) async fn fetch_pending_pairs(
     pool: &PgPool,
     target_exit_ids: Option<&[String]>,
 ) -> Result<Vec<PendingPair>, anyhow::Error> {
-    let qrows: Vec<(String, String, f64, f64, f64, f64, i32)> = match target_exit_ids {
+    let qrows: Vec<(String, String, f64, f64, f64, f64, f64)> = match target_exit_ids {
         Some(ids) => {
             sqlx::query_as(
                 "SELECT c.exit_id, c.place_id, \
